Targets are now tables containing name, path to binary and compatibility.
Targets now can be defined with target name, path to the target Lua
binary and the Lua version with which they are compatible.
For example, target "LuaJIT 2.0.3" is compatible with Lua 5.1.5,
so it could be defined as
{
  name = "LuaJIT 2.0.3",
  binary = path-to_luajit_binary,
  compatibility = "5.1.5",
}

Other attributes can be freely defined. These will be available to
task functions but ignored elsewhere. While not strictly
encouraged, this practice is convinient and can be used freely.

The order of targets in config is now explicitly preserved, instead of
implicit alphabetical order used until now.